The Global Institute of Hotel Management
The Global Institute of hotel management has been established to facilitate the transition from Hospitality Academics to successful careers in Hospitality. Hospitality, we believe, has new connotations in the modern world. It is an emerging industry with a bright future and part of the sunshine sector. Like always, we love to be at the vanguard of action and with a preparedness expected of the leaders in the industry. The horizons of the Indian hospitality industry continue to expand at a dynamic. Today the hospitality industry contributes invaluably to the Indian economy by creating employment and entrepreneurial opportunities and by being one of the highest grosser's of foreign exchange for the national treasury. Looking at the pace of growth of the hospitality industry, it is envisaged that there will soon be a marked shortage of managerial and technically skilled professionals. Keeping this in mind, our curriculum is designed to meet the challenges and demands of domestic and global markets.